Refactor: Phase 1A - Module separation and code organization
## 主な変更点

### モジュール構造の導入
- lib.rs: ライブラリエントリーポイントの作成
- app/: アプリケーションロジックの分離
  - error.rs: エラー型定義
  - mod.rs: App構造体とコア機能
- query/: クエリ処理の分離
  - mod.rs: JSONデータとjaqクエリ実行
- ui/: ユーザーインターフェースの分離
  - events.rs: イベント処理とAction定義
  - mod.rs: UI描画とターミナル管理

### コード改善
- main.rs: 716行 → 84行 (88%削減)
- 全体: 391行に分散構造化
- 責任の明確な分離
- テスト保持 (2 tests passed)

### 機能維持
- 既存の全機能が正常動作
- コンパイル・テスト・ビルド成功
- 後方互換性維持

この変更により、コードの可読性・保守性・拡張性が大幅に向上しました。
